Get the game working again after much bit rot. (#25)
 - serve frontend from App Engine to avoid extra bucket complexity and allow for https to "just work"
 - no longer using custom domain name
 - update js packages
 - update py packages
 - fix things after updates
 - tweaked default ports, add logging
 - fix bug in GameReducer.ts that would make some UI game updates not show until other events on those elements
 - clean up README.md s
 - personally recreated the project in Google Cloud, hopefully the $0.04/mo storage charge goes away

+++ scryfall/updateCards.py
@@ -33,7 +33,8 @@ if reload:
     with open(filename, mode="wb") as o:
         # json.dump(data1, o, ensure_ascii=False)  # some card names are unicode
         o.write(data1)
-    os.rename(default_cards_json, 'scryfall-default-cards.json.bk')
+    if os.path.exists(default_cards_json):
+        os.rename(default_cards_json, 'scryfall-default-cards.json.bk')
     shutil.copyfile(filename, default_cards_json)
     print("complete")
 else:
